+++ integrations/mistral/tests/test_mistral_chat_generator.py
@@ -96,7 +96,7 @@ class TestMistralChatGenerator:
 
     def test_init_fail_wo_api_key(self, monkeypatch):
         monkeypatch.delenv("MISTRAL_API_KEY", raising=False)
-        with pytest.raises(ValueError, match="None of the .* environment variables are set"):
+        with pytest.raises(ValueError, match=r"None of the .* environment variables are set"):
             MistralChatGenerator()
 
     def test_init_with_parameters(self):
@@ -191,7 +191,7 @@ class TestMistralChatGenerator:
                 "generation_kwargs": {"max_tokens": 10, "some_test_param": "test-params"},
             },
         }
-        with pytest.raises(ValueError, match="None of the .* environment variables are set"):
+        with pytest.raises(ValueError, match=r"None of the .* environment variables are set"):
             MistralChatGenerator.from_dict(data)
 
     def test_handle_stream_response(self):