+++ integrations/amazon_bedrock/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Changelog
 
+## [integrations/amazon_bedrock-v4.2.0] - 2025-09-19
+
+### 🚀 Features
+
+- Add a new `S3Downloader` component (#2192)
+
+## [integrations/amazon_bedrock-v4.1.0] - 2025-09-19
+
+### 🚀 Features
+
+- Support AWS Bedrock Guardrails in `AmazonBedrockChatGenerator` (#2284)
+
+### 🧹 Chores
+
+- Bedrock - remove unused `stop_words` init parameter (#2275)
+
+### 🌀 Miscellaneous
+
+- Chore: Fix linting aws bedrock (#2249)
+
 ## [integrations/amazon_bedrock-v4.0.0] - 2025-08-29
 
 ### 🚀 Features